requirements.txt
streamlit>=1.37
sqlmodel>=0.0.19
alembic>=1.13
psycopg2-binary>=2.9
duckdb>=0.10
duckdb-engine>=0.13
pydantic>=2.7
anthropic>=0.28
python-dotenv>=1.0
pytest>=8.0
pytest-cov>=5.0
black>=24.0
isort>=5.13
flake8>=7.0
interrogate>=1.7
pip-audit>=2.7
